ui: handle swapper task in multi-machine traces
This CL fixes the bug that swapper tasks are visible in multi-machine
traces, there there are multiple swapper UniqueTids, and the `utid != 0`
condition doesn't hold for multiple machines. The CL adds an `is_idle`
column to the thread table in the trace processor, and update queries
from `utid != 0` to:
1. `is_idle IS NULL` when referencing the thread table, or
2. `utid not in (select utid from thread where is_idle)` when
referencing other tables (e.g. sched).

Bug: 284258446
Test: Updated screenshot tests
Change-Id: I8bfda8b39e12b34bd06ec21a7a4cc1044a39b37a