ui: Remove the ability to set TrackNode ids manually
TrackNode ids were originally designed to be used for two purposes:
1. Mapping between DOM elements and TrackNode objects.
2. Allowing plugins to insert track nodes into groups defined by other
   plugins.

Since we now allow inter-plugin dependencies, we have no use for 2 any
more, which leaves 1 as the only reason for these ids to exist. Since
1 does not require plugins to be able to manually set these ids to well
known values, we can remove this feature and simply let the workspace
assign unique ids from an incrementing counter, which reduces the
likelihood of duplicate ids significantly.

This CL removes the option to define an id when creating a TrackNode,
and removes all usage.

Change-Id: I09ec63acbaa5c287121e97369b927b21988eac41