Import Bluetooth duration as dur rather than arg.
This imports the duration as the slice duration to show visually the
time period it covers and be more easily queryable. This patch also
changes the semantics of the proto to use nanosecond duration. The data
source is changed along with this, and is not yet widely used, so a
breaking change should be okay.

This change also adds the bluetooth events as one of the tracks shown in
the UI under System/Bluetooth.

Bug: 365778358
Test: run-dev-server and view trace
Change-Id: I0d39aea8dbd4c75634221c540b690ec3ba463c7c