ui: Add search implementation based on dataset
This search mechanism uses dataset and searches all tracks that have
ts, name, and id columns, (which includes debug tracks) globbing on the
name column.

This search mechanism is a WIP and is not feature complete compared to
the existing search mechanism, so it's behind a feature flag for now.

Change-Id: I6bfe654a4c3e6686a7606c95c40af781c90961bd