ui: Batch large union datasets into CTEs to avoid SQLite limits
SQLite has limits on the number of terms in a compound select statement.
By default this is set to 500, and TP is using the default.

There is a good reason for this limit. Compund select statements are
executed using a recursive algorithm, so we want to keep this limit
in place to avoid overflowing the stack.

See: https://www.sqlite.org/limits.html#max_compound_select

This CL adds a feature to UnionDataset which splits up the union all's
into batches of CTEs when the number of datasets in the union is
larger than 500.

In reality, we should never really hit these limits as queries should
be optimized to avoid it, but optimizations are always best-effort so
it's good to have this technique as a backup.

Change-Id: I3bbfcbb4b92577f8041e33a278a1664ddfcebd1c