Increase size of SMB when using shmem emulation
In unified tracing, during a tracing session the traced_probes on the
client side registers multiple data sources/trace writers. They concurrently try to acquire a chunk from the SMB. This causes the SMB
to fill up before all trace writers can acquire a chunk, which leads
to traced_probes to stall and crash because it cannot acquire a new
free chunk. Flushing in this case doesn't work because all the current
chunks are in "being written" state. To mitigate this we increased the
SMB size when in shmem emulation mode to 1MiB.

Test: tools/run_android_test --system-adb out/android perfetto_unittest
Bug: 399311463
Change-Id: I439c289759cb46659e5a322d178fc697d82e6787