traced_perf: add support for cpu masks
Supports both sampling and polling modes. It's added as a top-level perf
config option since while we already have a Scope proto nested inside
CallstackSampling, we need something that also applies to the polling
mode. In the future we'll likely unnest Scope to be top-level, but right
now that would be misleading, as most options are not applicable to
polling.

If no cpu mask is given, we default to all cpus as before.

Note that we do *not* automatically respect the
/sys/bus/event_source/.../cpumask, it's up to the user to figure out
which cpu values are legal for a given counter. An opportunity for a
future improvement.

Based on a patch by Nancy Zhao (zhaon@google.com).

Bug: 394723317
Change-Id: I6ef1a85016aacdb43bd00bb6f416c6ce713fb249