tp: perf_sample: remove legacy handling of counter-only packets
The removed early-exit comes from the older days, when all perf traces
were expected to have callstacks. Nowadays we properly support
counter-only modes, so we should just insert everything into the
perf_sample table.

Note on difftest: since callsite_id column is optional, the test query
wasn't accurate in the first place. It'd be nice to change the pprof
functions to handle absent callsites better, but that's outside of this
CL's scope.

Bug: 389550086
Change-Id: Ib8c0e13d12e8bfe962e5b70c6372c3df520e7f3b