tperf: emit all samples even when doing selective callstack sampling
Short version: fixes incorrect counter deltas during post processing.

Currently, we completely drop the samples if they don't match the
|callstack_sampling.scope| target filter, for example for kernel threads
when not recording kernel callstacks. This is a leftover from when the
profiler did not include the counter values themselves in the packets.
Now, these drops become gaps in the per-cpu counter timeseries, which
makes the counter tracks in the UI less accurate. More importantly,
if trace processing tries to assign the delta between two subsequent
samples to the latter callstack, we currently potentially assign the
incorrect delta if there were any dropped samples in-between the two
serialised ones.

Bug: 327359569
Change-Id: I8b4533be0d2afa4ab1df557f717bd74a3cd94838