stdlib: Account for negative swapper idle transition cost
In the idle transition cost function, account for the negative swapper
idle transition cost. The intent is that with the negative swapper idle
transition cost, the sum of every thread's idle transition costs is
equal to zero.

The idle transition cost function can then be used by the UI and metrics
to output the same data.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 403373348
Change-Id: I42fb9aa50df016ab0c93bd1ac22cbcef04d53c69
Signed-off-by: Samuel Wu <wusamuel@google.com>