Flush chunks before exceeding IPC buffer capacity in shmem emulation
When using shmem emulation, the completed chunks' data is passed through
the socket connection and handled in traced via the IPC frame
deserializer. Given that the IPC buffer has a set capacity, we need to
flush the completed chunks in the SMB early enough before the total
bytes of all the completed chunks surpass the IPC buffer capacity.

Hence, when a chunk is marked completed we immediately we trigger a
synchronous flush of pending commit data requests. This is only done
when using shmem emulation.

Test: tools/run_android_test --system-adb out/android perfetto_unittest
Bug: 399311463
Change-Id: I39e92b5ed63b3e1e90636c0b3c13bcbc31aadd7f