ui: Add virtually scrolled query results table
When users accidentially run a query that returns 10,000's of results,
forgetting to add 'limit X', this can produce an enormous results
table that can tank DOM rendering performance.

This CL adds a new implementation of the table which uses virtual
scrolling (based on the existing VirtualTable widget) which only
render the elements to the DOM which are actually within the scrollable
viewport.

This impl is behind a flag currently, as it lacks various features
from the current one.

Change-Id: I9cf4c3f71824fecdf956aa4e670d90034b59e63e