ui: Tidy up anchor element CSS to enable more elegant line wrapping
This CL reverts to using text-decoration for managing the underline of
Anchor widgets instead of border-bottom.

This is a much more elegant solution, looks a lot neater, and seems to
avoid any hacks put in to solve fiddly icon alignment, which fixes the
problem of anchors sometimes not wrapping onto multiple lines, and
looking ugly when they do.

Fixes: https://github.com/google/perfetto/issues/1037
Fixes: 404199404
Change-Id: Id21428c0440a3b2f82eaca06aaa87e1de5f35dc2