perfetto: remove mm_events support
If this broke you: please get in contact with us at
https://github.com/google/perfetto.

The kmem_activity work is not actively used or maintained, we haven't
heard anything about it since we asked last year. Removing the
associated code from the perfetto repo.

There's a public documentation page[1] and some scattered code references
across AOSP. I'll see what is easy to clean up, but the main priority is
to remove perfetto bits.

[1] https://source.android.com/docs/core/perf/mmevents-stats

Bug: 183037386
Change-Id: Ie8f1e99378c04be8b807a065ae44113d0271ce0a