sysprop ds: downgrade read errors to DLOG
We end up logging to logcat on every attempt to read unset/absent
sysprops. I am of the opinion that these logs are not useful outside of
debugging selinux permission errors, which should already be logged
separately.

DLOG means we won't log the errors (including on userdebug/eng), unless
the perfetto stack is rebuilt with verbose overrides.

Change-Id: Ie430e1ac76d73bc8efc5ccd1611da4a9ba1d991a