Fix BatteryStatsHistoryState parse error checking
when possible_history_state_item.ok() evaluates to false that means
we were not able to parse the history state item, however when that
happens we still return true, indicating a successful pars, which is
incorrect.

To fix this, return false when possible_history_state_item.ok() also
returns false.

As part of the fix, we remove the top-level "return true" so we no
longer fall through to this as a default in order to minimize the
likelyhood that similar oversights in the future result in incorrect
behavior.

Change-Id: I9e8b6f59c3bdf93e3bf8869a576689a210651229