ui: Improve peformance of heap graph context options
The previous implementation was designed for caching
in a table and post-filtering so it was inefficient
when used as a view. We now pre-filter the path hashes
before creating the view

Removed incoming_reference_count from the 'Objects' view
since the count is always one per object and we know the
number of objects.

Test: Manual
Change-Id: Ieaf9cebf9dbad4eadf4044ed68b5aaf35b17e887