fix single track counter aggegation query
There is a bug in the subqueries of the specialized single track
counter aggregation query.

They make refernce to ts & dur, but those refernces are ambiguous in
this context. "dur" for example, does not appear in the counter table
at all, meaning this value is likely being pulled from an arbitrary
row of the "res" CTE instead, which was not originaly intended.

In practice these incorrect subqueries have resulted in incorrect result
for first_value, which also impacts the Rate/s and delta values
presented to the user.

Out of the quick fixes I attemped, value_at_max_ts had the least negative
impact on perf, so went with that solution.

Bug: https://github.com/google/perfetto/issues/1050

Change-Id: I9a3a29145eddb1d3d921c2194c7fbd98051133da