Update the perfetto sdk to v49 (#145)
Updating just the sdk, and not the full perfetto repository.
The update is needed for the device generation of events starting
on different threads. With the old perfetto we did loose some
of the events.
Note that the perfetto is designed to be compatible in respect
to the used sdk.