ui: Fix heap graph plugin crash with multiple paths
The path hash is an aggregatable property, this means
that when we filter the flamegraph, any node can
have a comma separated list of path hashes. This list
was begin passed directly to the macro arg expecting
an Expr and of course crashed.

The current approach converts the comma separated
path_hash list to a table which is then passed as an arg
to the macro and evaluated as a JOIN. Using a view leads
to very slow joins. A table shouldn't consume much memory
because this is just the filtered list of paths to a given node.

Test: Manual
Change-Id: I7cfb6234a5addcc86f62e147a30f6f0221576812