ui: Added pluginable track filters
This CL adds the ability for plugins to add filter criteria, which
are additional ways to filter tracks.

Prime candidates are:
- filter by thread
- filter by process
- filter by cpu

For the purposes of demonstration and testing, this CL changes the
'Thread' plugin to add filter criteria for threads and processes.

This CL also restores the ability to pass multiple name filters separated by commas.

Change-Id: Id17150cddf38a1474c3715d9ffabe12fe84e86e5