Send guest VMs' SystemInfo packet to host traced
During traced_relay connection to the host traced instance, we send the
SystemInfo packet of the guest VM. Once received, traced caches the
serialized system info and emits it to the trace file at the end of a
tracing session.

Having the SystemInfo packet of all secondary VMs allows us to do
syscall id mapping to syscall name and also provides useful information
to understand all the machines involved during a tracing session.

Test: tools/run_android_test --system-adb out/android perfetto_unittest
Bug: 402546586
Change-Id: Icda20a44d3e2f402e979031eb836720e7da3ff3a