ui: Use specialization to fix parsing bug
The current lezer SQL parser has a bug where it misidentifies
identifiers with keywords in them.

E.g. consider the identifier 'COUNTER'.

COUNTER
'---'''
  |   `-- ER - Identifier
   `----- COUNT - Keyword

The current incarnation identifiers this as a keyword followed by an
identifier, so the syntax highlighting is wrong.

Turns out lezer has a tools for this exact problem - specialization.

https://lezer.codemirror.net/docs/guide/#token-specialization

This CL uses this tool to make it so that COUNTER is identified
entirely as an identifier.

Change-Id: Ifb297ae28866437a8c4891ebda2b62a48867e212