ui: Add MultiselectInput widget
This widget works functionally the same as the Multiselect widget
but provides the ability to select zero or more items from a list by
typing the labels into a text box instead of selecting from a list.

While typing, autocomplete suggestions appear in a popup below the
input field, and selected items appear as tags on the left hand side
of the input.

This input can provide a more keyboard-driven way to enter multiselect
options with fewer clicks.

Change-Id: Ie94e3fa926b54f53e3a9716cd54bd98ea485296b