tp: make public _slice_with_thread_and_process_info
This CL makes a widely used private table of TP public and publishes it
in the right place.

PSA: if this CL broke you, it's a warning to NOT rely on private tables
of TP. Please reach out to us to make tables public or you are liable
for breakages like this.

Change-Id: Ibef6ebb5de8af939a4d4f49491604a574c9c4878