perfetto cts: split ART's MTS test cases out of CtsPerfettoTestCases, into a new target called CtsHeapprofdJavaCtsTest
The current configuration is incorrect in the context of MCTS, because
the entirety of CtsPerfettoTestCases is being run on older platforms.
The new platform-oriented tests (e.g. the behaviour of /system binaries)
will never succeed there since the associated platform code is absent.

This is because the MTS->MCTS migration overlooked the fact that ART MTS
ran only a subset of targets in the GTest CtsPerfettoTestCases binary[1].

This patch splits the MTS-specific tests into a new test target. Note
that it also introduces two new tests into the ART MTS. Since the custom
allocator NDK API tests were not matching the test name filter[1].

The allocator test were called:
* HeapprofdCtsTest::JavaHeapRuntime
* HeapprofdCtsTest::JavaHeapStartup
And they are now called:
* HeapprofdJavaCtsTest::ArtHeapCustomAllocatorRuntime
* HeapprofdJavaCtsTest::ArtHeapCustomAllocatorStartup

More details on the bug.

[1] https://cs.android.com/android/platform/superproject/main/+/main:test/mts/tools/mts-tradefed/res/config/mts-art-extra-tests-list-user.xml;l=17;drc=be9d63305cffabfea54646f7a2490f3d37e620e8

Tested: atest CtsPerfettoTestCases;
Tested: atest CtsHeapprofdJavaCtsTest
Tested: mts-tradefed run commandAndExit mts-art-extra
Bug: 406239886
Change-Id: I5170d2aca09adc6c908e3813fed1574256f8a6a6