tprobes: debuggable android builds: optionally serialise bad pages
To assist internal debugging. The best way would've been to retain the
page until end-of-trace, as otherwise we'll miss cases that rotate out
of ring buffer modes. However that proved to be too much plumbing for a
debugging feature (cpu_reader -> OnFtraceWritten -> FtraceDataSource, or
static storage + accessors).

I can be convinced to change things, but this feels like the minimal
addition.

Bug: 401855064