protos: remove unused import
Oddly, nothing in our tools/build catches this. I couldn't find an easy
way to get this out of protoc (playing around with the invocation in
tools/gen_binary_descriptors). So just removing the import for now, and
making sure it's covered by presubmit is a task for another day.

Bug: 406928883