stdlib: Exclude dur==0 in Wattson freq/idle tables
Exclude slices with duration of 0 in the initial freq and idle tables
used during Wattson calculation. Removing the dur==0 slices is necessary
since a requirement for interval_intersect() is that the input have
non-zero slices duration.

A dur==0 freq or idle slice can occur in the rare case when a trace has
a freq or idle ftrace event at the same timestamp as trace_end().

Bug: 408009977
Test: tools/diff_test_trace_processor.py out/linux_clang_release/trace_processor_shell --name-filter '.*wattson.*'
Signed-off-by: Samuel Wu <wusamuel@google.com>