df: add runtime builder class for building dataframes for unknown schemas
This CL adds a builder class for building dataframes where the schema of the
dataframe is not known upfront and should be inferred from the inserted
rows. We try and always be as memory efficient as possible and infer as many
properties (i.e. sorted/id etc) as we can.