tp: make sure all virtual table modules receive Commit/Rollback hooks (#1176)
This CL ensures that the commit/rollback hooks is plumbed through to
every virtual table module in trace processor, not just the ones which
are considered part of the engine.

To do this, we needed to rework a little bit how the ModuleStateManager
works to be an object you inherit from instead of one you put in your
context. We also need to extract out a base class from
ModuleStateManager so the engine can keep a hold of the type erased
pointers (with the added benefit of less binary size).

In the process, we remove several hacky use of sqlite_engine() now that
RegisterVirtualTableModule is directly plumbed into the
PerfettoSqlEngine.