Update TransitionsTracker to use own ArgsTracker instances (#1195)
This is to fix an issue where ShellTransition data was not being
correctly parsed by the parser resulting in missing data in the
transitions in Flicker and Winscope. The was due to a misuse of the
ArgsTracker interface in the ShellTransitionsParser, where the args were
being overwritten with new arg sets due to the flushing of the
`context->args_tracker` that was happening every once in a while. We
should now be accumulating the args correctly and fully before flushing
them to storage, ensuring we don't miss any args in the trace tables.

Bug: 406973318
Test: tools/diff_test_trace_processor.py
out/linux_clang_debug/trace_processor_shell
--name-filter=".*Transition.*"
Change-Id: I74728cb062066468dd80f5aa6315e7e15316343e