tprobes: generic events as self-describing protos (2/3)
The encoding of ftrace events that do not have a protobuf description in
the perfetto repo is very space inefficient. This is because every
GenericFtraceEvent proto has a string description of the field names and
added layers of proto nesting. This was good enough for local
exploration, but is often used by accident when configs use category
globs. Also, long-term it'd be nice to move away from requiring protos
to be checked into our repo outside of truly well-known cases.

This patch adds a feature flag to enable such generic events to be
serialised using self-describing protos, with the aim of being a bit
more space efficient. The FtraceEvent proto will have a submessage under
a field id not covered by the compile-time oneof fields (similar to
proto extensions). The FtraceEventBundle will have an
id->ProtoDescriptor map, once per generic event type.

Notes:
* every unknown event type will have its own proto field id, and a
  bundle-level proto descriptor.
* the mapping stays consistent within all bundles in  a tracing session.
* there is no global mapping of event->id across traces.
* we still use the FtraceEvent container proto, with the upsides of
  being able to do streaming serialisation of bundles, and the downsides
  of protobuf encoding overheads due to deep nesting (and extra missing
  bits like delta timestamps).
* we do not use incremental state, every bundle describes the generic
  events. The space cost is amortised by bundles being relatively
  infrequent.
* the new encoding is behind a flag, and the default case continues to
  be GenericFtraceEvent protos.

//////////////
For reviewers:

This might help visualise the proto structure:
  message FtraceEvent {
    optional uint64 timestamp = 1;
    optional uint32 pid = 2;

    oneof event {
      PrintFtraceEvent print = 3;
      SchedSwitchFtraceEvent sched_switch = 4;
      ...
    }

    // synthesised at runtime, similar to an extension range:
    optional GenericOne unknown_event_one = 65536;
    optional GenericTwo unknown_event_two = 65537;
  }

I started by reserving a high range of field ids within the FtraceEvent
proto for the synthesised-on-the-fly submesssages.  Would've been nice
to use something with one fewer byte for pb tags (e.g. 1k-2k), but that
is a bit too close to the compile-time proto range, which already goes
up to ~600. The field id range does matter since we use "id>65k" as a
way of quickly identifying that an event is generic within the
proto_translation_table and cpu_reader.

proto_translation_table: I changed the |events_| mapping (from ftrace id
to a struct describing the event) to generate proper per-field proto
descriptions of events, and moved the GenericFtraceEvent specifics to
cpu_reader's serialisation path.

cpu_reader::Bundler changes: when parsing, we need to remember the set
of referenced generic events to emit their descriptors.