Allow users of DatasetSliceTrack to entirely customize what's shown on the tooltip (#1232)
Users of DatasetSliceTrack can customize what's shown on the tooltip by setting the tooltip attribute. This should be a function which is called every render cycle, and it should return the mithril vnodes to render on the tooltip.

This PR also stylizes the tooltip, putting the duration in bold first followed by the title.

It also modifies TraceProcessorTrack to add the category field on the tooltip if non-null.