Tidy up page registration (#1265)
In order to register a page plugins currently need to provide a
component whose attributes extend the `PageWithTraceAttrs` interface,
unless the magic `traceless` flag is passed at which point they need
only pass a component whose attributes extend the `PageAttrs` interface.

This is confusing for a number of reasons. How is the trace passed in to
the page? What about the app? How do I pass my own arguments into my
page?

Consider tracks and tabs for example, Any Mithril rendering done by
those entities is described using functions, which are simply called
every render cycle (when relevant). Usage of a function allows crucial
arguments such as trace/app/etc to be captured using a closure, and any
'this' context to be captured simply by using an arrow function.

The page API, however, takes a Mithril component rather than a function,
which makes it very unclear on how to inject other arguments, and
mysteriously hides how or if the trace is injected. Oh and if you get
the attributes type wrong on your component, good luck de cyphering the
TS error message.

Anna and Lydia both already got immediately stung by the issue of
injecting extra arguments form their plugin into the page component when
working on the explore page, so this is a real problem. The provided
`bindMithrilAttrs` function, which was supposed to solve this problem,
doesn't help as it's a very confusing templated function which, combined
with Mithril's uneasy TypeScript integration, just makes the situation
even worse.

I'd argue that one of the biggest issues is that it's just inconsistent
with the rest of the other APIs, causing unnecessary friction when
trying to write plugins.

In addition it pushes the onus of pulling out plugin-contextualized
trace context objects onto the router, which is redundant as we have
already solved the problem elegantly several times in the other APIs
without those subsystems having to know anything about plugins - by just
having the plugin inject in the trace object that was passed to it in
its onTraceLoad hook.

Finally, this architecture promotes the proliferation of the pointless
PageWithTraceAttrs, PageWithTraceImplAttrs interfaces which don't really
prevent any boilerplate and just add unnecessary abstraction and
coupling. When using a Mithril component, there is an unwritten
convention that its attributes interface is listed in full above it so
we can easily see how to use the component. The proliferation of
PageWithTraceAttrs and friends just hides a couple of these attributes,
and makes the code confusing to read.

This PR fixes things up by changing the page subsystem to take a render
function instead of a component. This makes the API much cleaner and
more consistent with the rest of the API.

New API:
```ts
trace.pages.registerPage({
  route: '/foo',

  // The render function is passed the subpage from the router, which can always be ignored
  render: (subpage) => m(FooPage, {subpage, trace, foo, bar, baz}),
});
```

Changes:
- Removed PageAttrs and PageWithTraceAttrs interfaces to simplify page
attribute management.
- Updated page registration to use render functions instead of page
components directly.
- Adjusted various page components to align with the new attribute
structure.
- Reduce coupling and enhance code readability by defining trace
attributes where necessary.