Add track descriptions (#1325)
This PR adds track descriptions to track definitions.

Plugins can attach descriptions (short strings of text) that are
displayed in the track details popup.

In the future we could make this style of help text more visible and
more consistent throughout the UI, but for now this change provides the
plumbing to allow plugins to specify descriptions.


![image](https://github.com/user-attachments/assets/2eafd89b-41c2-4fdf-a1a2-7c68b246f253)