frozen_ftrace: Make event_format_path optional
event_format_path is required for analyzing the dynamic events like
kprobes, or module events which is dynamically added. But for the
normal usage, the static event ids should not be changed across the
reboot.

Thus I decided to make it optional and simplified the perfetto.rc by
skipping event directory backup processes.

Change-Id: I773554f999f12232dfcba309cf6c30bcf1587a7c