build: don't use gold on Linux standalone builds (#1350)
On Linux, when NOT using clang, we used to depend on the gold linker.
That linker is now going away, making this problematic. We realistically
don't care about it, because all our reference builds (both on Android
and linux) use the hermetic clang toolchain.

I tested that this change has only effect on the gcc build configs, as
follows:
- Run tools/setup_all_configs.py --android
- cp -a out out.old
- Apply this patch
- cd out; for x in *; do gn gen $x; done
- git diff --no-index --word-diff=color --minimal out.old out

The only changes in ninja files are in the out/linux_gcc* dirs.  

Bug: b/410777878