Remove `python_version` from python/BUILD file (#1365)
The copybara worker is blocked on wrong formatting of the python/BUILD
file: we can't mention `python_version` anymore and the LSC is removing
it.

Editing gen_bazel to stop adding it to autogenerated python BUILD file.

tools/install_build_deps install bazel with version 7.4. I built it
locally and it seemed to build fine