Update index.ts
Complete perfetto internal systemui name

Because the underlying AndroidRuntime::setArgv0 will truncate thread names longer than 15 to keep 15, so this is probably intentional. Because the underlying setArgv0 is consistent with the kernel thread name length, this modification is to keep the first letter capitalized with other variables in TRACKS_TO_PIN, and to make the code clearer after completion, and easier for users to understand.