Add machine id column to AndroidLog panel
In multi-VM traces, it is valuable to differentiate from which machine
each Android log is coming from. For this we added a machine id column
to the table in the AndroidLog panel. This columns is conditional to
there being more than one machine, if not then the log panel behaves the
same way as before.

Test: ui/run-unittests
Bug: 406061688