[ui] Add improved bar chart to the table viewer. (#1347)
The new bar chart fetches the necessary values itself from the SQL
instead of relying all of rows to be passed to it, which makes it much
more suitable for large data-sets.

It also shows a popup on click, showing the user the % figure of the
value and allows adding filters.

This patch is based on
https://android-review.googlesource.com/c/platform/external/perfetto/+/3517372.

Bug: 395565690