Add machine id filter to AndroidLog (#1389)
In multi-machine traces doing log analysis might be complicated given
there could be thousands of logs from each machine in the same log
table. To improve the log navigation, we added a multi-select filter
which allows the user to select any subset of machines to only view
their respective logs.

Test: ui/run-unittests
Bug: 406061688