Replace the deprecated onMessageRefresh with present
Remove the abandoned onMessageRefresh in old Android versions

In Android T and higher, the atrace event of SurfaceFlinger::onMessageRefresh() is indeed deprecated, so the logic of these test codes needs to be adjusted. If the test is only for Android T+, it is recommended to directly delete all onMessageRefresh related Trace events, because the function has been removed from the code base and its Trace events are no longer generated. In addition, the function of onMessageRefresh has been refactored and optimized in Android T, and its core logic has been integrated into the new process of CompositionEngine. It is recommended to replace onMessageRefresh here with the Trace event of the new function (such as CompositionEngine::present()).