Fix bug in heap profile auto-selection on trace load
When loading a trace with heap profile tracks, we automatically select
the earliest heap profile sample on trace load.

We currently assume that this event has id = 0, but in reality this is
not necessarily the case.

This change modifies the heap profile plugin so that it works out the
event id properly.

This change also refactors the selection code a little to make the
error message clearer and to ease debugging in the future.

Change-Id: I62c3e62243505426ba5b7390b6a5a30d7caef161