Remove `utid` field from core selection object (#1434)
This patch removes the `utid` field from the core selection object.

The only component that uses this field is the critical path plugin.
This has been changed to look up the utid on demand from the track tags
and/or the dataset.

This is part of the movement to remove anything thread related from the
core.