Fix crash in heap profile auto-selection at trace load time (#1435)
https://buganizer.corp.google.com/issues/416190258

When loading a trace with heap profile tracks, we automatically select
the earliest heap profile sample on trace load.
    
We currently assume that this event has id = 0, but this is not
necessarily the case.
    
This change refactors the queries in the heap profile plugin to make
things a lot clearer, as it was getting out of control, and also fixes
the bug so that it works out the necessary event id properly.
    
This change also refactors the selection logic in SelectionManager a
little to make the error message clearer and to ease debugging in the
future.