Perfetto: Add timer ftrace events (#1472)
In debugging issues around timer expiration, I found collecting
'timer/*' events in the config resulted in ftrace entries in
the ftrace track, however for hrtimer_expire_entry() the event
had  just an opaque index number for the function argument
instead of the function name, limiting its use in debugging.
    
Thus, this series adds some logic to the ftrace_parser so the
function name is preserved in the ftrace entry.
    
Additionally, it parses the hrtimer_expire_entry/exit events
to add slices to a new HRTimer track in the UI (similar to
IRQ and SOFTIRQ tracks), so that one can visualize when the
timer expiration happened.

---------

Signed-off-by: John Stultz <jstultz@google.com>