Fix android_power_rails_counters average power calculation (#1455)
"delta" value is defined as the elapsed energy since the PREVIOUS
sample, however "dur" is defined as the elapsed time until the NEXT
sample.

Whether we choose to use the next, or previous sample for this
calculation, we should make sure we use the same for both the delta
energy and delta time values.

Since the documentation indicated that average_power it should be the
next sample, adjust the query to ensure we are using both delta energy
and delta time relative to the next sample.

Since the sample period for energy samples is fairly uniform this
doesn't impact the power values too much but does shift power values
later by one sample period.